+++ js/syllabus.js
@@ -2941,3 +2941,4 @@ loadEdexcelSyllabus();
 loadHHKSyllabus();
 /* Pre-load knowledge point data for active boards */
 loadKnowledgeData('cie');
+loadKnowledgeData('hhk');